[TensorPipe] Implement join correctly
My understanding of how the RPC shutdown works is as follows:
- All workers enter a barrier (i.e., `_wait_all_workers`) so that "passive" workers (the ones that don't initiate any RPC call themselves, but just receive and handle requests from other workers) stay alive until the "active" workers have also completed. This way we know that when the barrier returns all workers are ready to shut down.
- Then we start shutting down the layers of the stack that are between the user and the "core" RPC agent, namely the dist autograd context and the rref context. These cleanups trigger some asynchronous RPC calls but, apparently, do *not* wait for their futures to complete. (This is because if they did then this would also cause some operations to become blocking during normal operations).
- Then we call `join` on the RPC worker (more on this later).
- Then we do another barrier.
- Then we shutdown the agent (abruptly, aborting any pending operation).

From the above, I believe the only requirement that is put on `join` in order for the shutdown to be graceful (i.e., for no operation to actually be aborted) is for `join` to wait for all *outgoing* requests to finish or, equivalently, for all pending futures to be completed. In other words, the "client" part of the agent needs to flush. I don't think we need to worry about the "server" part of the agent, i.e., we do not need to worry about waiting for the requests the agent is currently executing on behalf of other workers, because we will implicitly wait for them in the barrier that follows the `join`.

Based on this understanding, in this diff I am changing the `join` method of the TensorPipe agent to wait until `clientActiveCalls_` reaches zero.

Differential Revision: [D21703020](https://our.internmc.facebook.com/intern/diff/D21703020/)

ghstack-source-id: 104584466
Pull Request resolved: https://github.com/pytorch/pytorch/pull/38933