Add interpolate-style overloads to aten::upsample* ops
Pull Request resolved: https://github.com/pytorch/pytorch/pull/37176

The non-deprecated user-facing interface to these ops (F.interpolate)
has a good interface: output size and scale are both specified as
a scalar or list, and exactly one must be present.  These aten ops
have an older, clunkier interface where output size is required and
scales are specified as separate optional scalars per dimension.

This change adds new overloads to the aten ops that match the interface
of interpolate.  The plan is to eventually remove the old overloads,
resulting in roughly net-zero code added.  I also believe it is possible
to push this interface down further, eliminating multiple optional<double>
arguments, and simplifying the implementations.

The rollout plan is to land this, wait for a reasonable interval for
forwards-compatibility (maybe 1 week?), land the change that updates
interpolate to call these overloads, wait for a reasonable interval
for backwards-compatibility (maybe 6 months?), then remove the old
overloads.

Note that this initial revision only updates upsample_nearest2d
(9 C++ functions).  I want to get agreement on the implementation
before I do the other 50.
ghstack-source-id: 104586366

Differential Revision: [D21209989](https://our.internmc.facebook.com/intern/diff/D21209989/)